requirements.txt
fastapi
uvicorn
python-multipart
websockets
pyperclip
Pillow
qrcode[pil]
websocket-client
python-dotenv
pywin32; sys_platform == "win32"
requests
